+++ src/examples/gsheets/dynamic/dynamic_flow.py
@@ -4,7 +4,7 @@ from prefect.futures import wait
 from examples.gsheets.utils import task_deploy_primitive_if_needed, task_filter_by_source_id, task_normalize_source, task_prepare_records_for_tsn
 from tsn_adapters.tasks.github import task_read_repo_csv_file
 from tsn_adapters.tasks.gsheet import task_read_gsheet
-from tsn_adapters.tasks.tsn import task_insert_tsn_records, task_get_all_tsn_records
+from tsn_adapters.tasks.trufnetwork import task_insert_tsn_records, task_get_all_tsn_records
 from tsn_adapters.tasks.data_manipulation import task_reconcile_data
 import trufnetwork_sdk_py.client as tn_client
