+++ src/tsn_adapters/tasks/argentina/models/aggregated_prices.py
@@ -5,8 +5,10 @@ This module contains the Pandera model for validating and typing the aggregated
 price data from the SEPA dataset.
 """
 
-from pandera.typing import Series
+from pandera.typing import Series, DataFrame
 from pandera import DataFrameModel
+from tsn_adapters.tasks.trufnetwork.models import TnRecordModel
+import pandera as pa
 
 
 class SepaAggregatedPricesModel(DataFrameModel):
@@ -30,5 +32,13 @@ class SepaAggregatedPricesModel(DataFrameModel):
     date: Series[str]
 
     class Config:
-        strict = True
+        strict = 'filter'
         coerce = True 
+
+
+@pa.check_types
+def sepa_aggregated_prices_to_tn_records(df: DataFrame[SepaAggregatedPricesModel]) -> DataFrame[TnRecordModel]:
+    new_df = df.copy()
+    new_df['value'] = new_df['avg_price'].astype(str)
+    new_df['date'] = new_df['date'].astype(str)
+    return DataFrame[TnRecordModel](new_df)
