+++ src/tsn_adapters/tasks/trufnetwork/models/__init__.py
@@ -1,0 +1,4 @@
+from .tn_models import TnRecordModel
+
+
+__all__ = ["TnRecordModel"]
