+++ src/tsn_adapters/tasks/trufnetwork/tn.py
